web/requirements.txt
python-dotenv
flask
flask-socketio
flask-admin
flask-login
redis
python-redis-lock
bcrypt
peewee
wtf-peewee
eventlet
redis
python-redis-lock
